Return error when loading fails instead of panicking
This refactoring changes the public API of the `openvino` crate to add a new error type, `LoadingError`, as well as a wrapper type, `SetupError`, to disambiguate between that and `InferenceError`. The `Core::new` function changes to return the wrapper, `SetupError`, since it could fail while loading the OpenVINO shared libraries.