Add strict clippy scanning to crates
This change configures rather pedantic settings for `cargo clippy`.
These crates, as bindings to an underlying API, use a limited subset of
the language so very few settings need to be ignored. If the
`clippy::pedantic` warnings become burdensome in the future, this can be
removed, but this commit makes the necessary changes to run without
warnings.