Upgrade to use upstream OpenVINO 2022.1 release
The 2022 release changes the names of several of the shared libraries
produced by OpenVINO as well as their directory structure. The C API is
still available, however, so this change continues to use that for the
time being instead of the newer v2.0 API.

Because of the significant changes to the release structure, there is a
decision to be made regarding backwards compatibility--e.g., using older
versions of OpenVINO with this binding crate. This change makes the
decision to stop supporting pre-2022 installations of OpenVINO in
ongoing versions of this crate: once this crate is released (`v0.4.0`),
users will have to decide whether to use this version for the newest
versions of OpenVINO or downgrade to a previously-released crate
(`v0.3.*`) for pre-2022 versions of OpenVINO.