Remove the `from-source` feature (#39)
Prior to this change, the `openvino-sys` crate had the ability to build
OpenVINO from source using CMake from within its `build.rs` script. This
increased complexity in various places (documentation, `Cargo.toml`
features, `build.rs`) and was prone to breakage as OpenVINO modified its
available CMake build features.

Instead of providing a `from-source` feature to do all this, this change
documents an `OPENVINO_BUILD_DIR` environment variable that can be used
for the same purpose: once a user has built OpenVINO from source on
their system, they can provide the path to this repository for the
`openvino-sys` crate to bind to. The `openvino-finder` crate contains
the necessary paths to find the default locations of all of required
libraries.