Double-check size of returned blob dimensions
As discussed [here], this change adds a (perhaps overkill) safety check
of the size of the dimensions array returned by the OpenVINO library.
The added assertion is trying to avoid the improbable case where some
future version of the OpenVINO library returns a dimensions array with
size different than the one auto-generated in the bindings; see `struct
dimensions` in `openvino-sys/src/generated/types.rs`.

[here]: https://github.com/intel/openvino-rs/pull/56#discussion_r1041570646