openvino-finder: locate `plugins.xml`
This change adds a `find_plugins_xml` function to `openvino-finder` to
locate the XML configuration describing the plugin libraries. In
previous versions, the `plugins.xml` file was in the same directory as
the shared library; in v2022.3, however, APT installations of OpenVINO
have the `plugins.xml` file in a versioned sub-directory (e.g.,
`openvino-2022.3.0/`). When no path is provided to `Core::new`, this
change will now search for the latest library and discover the
`plugins.xml` file based on that. There are several escape hatches for
specifying the path directly: pass the path directly to `Core::new` or
set the `OPENVINO_PLUGINS_XML` environment variable.