Update crate dependencies (#69)
This is not only regular dependency gardening, it also switches from
`structopt` to `clap` to avoid some unmaintained dependencies (`atty`,
`ansi_term`). This also switches to `env_logger` instead of the
`pretty_env_logger` wrapper.

As a result of updating `bindgen`, the bindings need to be updated
(only minor tweaks) and the `libloading` library must be ignored
by cargo-deny (`bindgen` has a different version than the
`openvino-sys` crate).