Handle "plugin not found" cases more gracefully
OpenVINO can be configured with a `plugins.xml` file that points to
libraries for running inferences on other devices. Previously, if this
file was not provided and `openvino-finder` could not find it, the
bindings crate would fail with an error. But this is not always
necessary: if the plugin libraries are located in the same directory as
the OpenVINO library, then a `plugins.xml` file is not needed. This is
the case for APT-installed versions of the package, e.g. In these cases,
we can pass an empty string to inform OpenVINO to look in the same
directory as the main library ([example]).

[example]: https://github.com/openvinotoolkit/openvino/blob/05a24b17769cb634f16343d551e25f542f0855b1/docs/snippets/ie_common.c#L14