ci: test more versions on more platforms
This change completely refactors the "build and test" story for CI. With
some unfortunate omissions (MacOS entirely, Windows dynamic linking),
the bindings are checked against the last three versions of OpenVINO on
more OS combinations. This is possible due to the new capabilities of
the `install-openvino-action`, which downloads an OpenVINO archive and
can now setup the runner environment. More work is needed, but this is a
big step in the right direction.

Because `install-openvino-action` is used more extensively, the previous
"build and test" job using Docker is no longer needed. The Dockerfile is
removed.

Also, since the "build from source" option is not tested in CI, we avoid
checking out all of OpenVINO's submodules, speeding up the build. Only a
few C headers are necessary for the binding regeneration, IIRC.