chore: update OpenCV version in `openvino-tensor-converter`
The `openvino-tensor-converter` crate is a helpful extension to this
project but not crucial, thus it does not receive as much attention as
other parts. It is useful for generating tensor files of various shapes
from images (e.g., for test fixtures), so I haven't felt the need to
remove it yet. Its dependencies had fallen far out of date and included
a very old version of `shlex` that has security vulnerabilities. By
updating to the latest `opencv` crate version, this dependency is also
updated. This change also updates the converter code to fit the latest
`opencv` API shape and `cargo-clippy` guidance.