Refactor the pre-processing pipeline (#103)
This change refactors the `prepostprocess` module by moving code around,
renaming things, and eliminating some extra code. The one key change to
look out for that is not trivial is the switch to use `&mut self` for
functions that use `*mut ...` pointers. This makes use of Rust's builtin
borrow checking to avoid concurrent mutable access to the OpenVINO
internal objects, which is hidden away by our `unsafe` FFI calls.

Co-authored-by: Rahul <rahul.s.chaphalkar@intel.com>