Use `slice::align_to`
When checking that tensor data can be converted to a slice of `T`, we
previously were checking the right things: is the length right? Is the
alignment correct? Turns out there is a `std` function that does this
for us: `slice::align_to`. Replacing the custom check with the `std`
version should have no effect on the code other than clarity.