Generate C enums as Rust enums
This change is an effort avoid missing new enum variants if they are
added to OpenVINO's C headers. As it stands now, new variants could be
missed until runtime, when the enum code can't be parsed and the library
panics. Instead, this change generates a Rust `enum` for each C `enum`
so that missing variants become a compile time error; if we update the
bindings but don't handle these, we see a compile-time error.