Propagate new `ov_*_e` enum changes throughout
The `openvino` crate depended on the previous constants generated by
`bindgen`. This change uses the new Rust `enum`s instead, creating
`From` and `Into` implementations to go between the Rust and C versions.

Why not just pass through the `openvino_sys` enum as-is? The only real
issue currently is documentation: the C headers start documentation with
`<!` which seems a bit weird. I've also fixed up a few grammar problems
in the documentation. But one day we may indeed be able to just reuse
these types.