Temporarily remove documentation for `ov_element_type_e_OV_BOOLEAN`
It's strange, but some combination of `bindgen` and `clang` does not see
fit to print the documentation for the boolean element type; I suspect
this is because it contains the word "boolean" but I'm not sure.
Perusing some Docker images, I found that `clang` v18 does not have this
problem when the code is re-generated. But earlier versions, like
GitHub's default `clang` v14 (and v16) want to remove this line. Let's
remove this line for now until the GitHub runners catch up.