Update bindings to v2024.4.0 (#144)
* Update bindings to use OpenVINO v2024.4.0

This consists of using #142:

```
$ cargo xtask update 2024.4.0
$ cargo xtask codegen
$ cargo test
```

The follow-on commits resolve the errors introduced by this update.

* Generate C enums as Rust enums

This change is an effort avoid missing new enum variants if they are
added to OpenVINO's C headers. As it stands now, new variants could be
missed until runtime, when the enum code can't be parsed and the library
panics. Instead, this change generates a Rust `enum` for each C `enum`
so that missing variants become a compile time error; if we update the
bindings but don't handle these, we see a compile-time error.

* Propagate new `ov_*_e` enum changes throughout

The `openvino` crate depended on the previous constants generated by
`bindgen`. This change uses the new Rust `enum`s instead, creating
`From` and `Into` implementations to go between the Rust and C versions.

Why not just pass through the `openvino_sys` enum as-is? The only real
issue currently is documentation: the C headers start documentation with
`<!` which seems a bit weird. I've also fixed up a few grammar problems
in the documentation. But one day we may indeed be able to just reuse
these types.

* Move `version` into its module; add parsing

It wasn't clear to me why the function and the struct were kept in
different places so I moved them together. I also added a
`Version::parts` function.

* Handle tests which depend on v2024.2 breaking change

Any test using `ElementType::U8` will cause problems when run on older
versions of the library. This change manually ignores tests that depend
on this by detecting the library version. For `Tensor`'s doc test, I
chose to use a type not affected by the breaking change to keep things
consistent.

* ci: update versions tested against

* clippy: fix `Into` to `From` recommendations

* fix: tweak version check

* fix: continue ignoring `OV_BOOLEAN` doc comment

See b2bdfb0c7 for the rationale for this.

* review: print message when test is skipped