Check version; fail if pre-2024.2 (#149)
* Check version; fail if pre-2024.2

In #143, we discovered that upstream OpenVINO had changed the
`ov_element_type_e` enum in a backwards-incompatible way. This means
that a mismatch between the Rust bindings (based on the C headers) and
the shared library will result in tensors being created with the wrong
type (e.g., `U64` instead of `U8`). To avoid this situation, this change
reads the OpenVINO library version at runtime (e.g., every time a `Core`
is created) and fails with an error if the version is older than 2024.2,
when the breaking change was introduced.

The effect of merging this change would be that newer bindings will fail
when used with older libraries, which could be problematic for users.
Users could always use an older version of the library (e.g., `v0.7.*`)
instead. And the alternative, letting users create tensors of the wrong
type, seems like it would open up the door to a slew of reported issues.

* Add documentation to helper functions

* Update CI versions: library + OS

This updates CI to test only the latest three OpenVINO versions, since
older versions will no longer be accessible due to the new "breaking
change version check." It also updates the OS version of the GitHub
runner for good measure.

* ci: specify ubuntu-24.04

* ci: roll back OS version

OpenVINO has not published packages for `ubuntu-24.04`.