Upgrade several dependencies
This PR used `cargo upgrade` from the `cargo-edit` crate to upgrade
several dependencies in this project. One notable absence is `bindgen`,
which `cargo upgrade` says it could not upgrade in a compatible way
(this is probably the right thing for now since it would require tweaks
to the generated files).

This upgrade moves the version of `anstream` to 0.6.18 which resolves
the unsoundness reported in [RUSTSEC-2024-0404]. `anstream` is used by
the `xtask` crate (not public; only for local developer tooling) and as
a test dependency of `openvino-sys`, so the unsoundness does not affect
the published library crates.

[RUSTSEC-2024-0404]: https://rustsec.org/advisories/RUSTSEC-2024-0404