Make code scan more robust (#158)
This CI check fails if `cargo clippy` fails, which is unfortunate since
we really want the output of `cargo clippy`, errors and all, available
to the CodeQL action. This change attempts to remedy that by ignoring
any error code returned by `cargo clippy` (and only `cargo clippy`; we
expect the other steps to succeed.