Upgrade bindings to v2025.1.0 (#168)
* Update `upstream` submodule to v2025.1.0

This captures (minus some cleanup of now-untracked files):

```
cd crates/openvino-sys/upstream
git checkout 2025.1.0
git submodule update --init --recursive
```

This brings the upstream submodule in line with release [2025.1.0].

[2025.1.0]: https://github.com/openvinotoolkit/openvino/releases/tag/2025.1.0

* Update bindings

This captures:

```
cargo xtask codegen
```

The key change introduced here--a breaking one--is the rearranging of
all of the enum values for `ov_element_type_e`. Like #143, this creates
an incompatibility when running a pre-2025.1 OpenVINO with this version
of the bindings.

* Remove `RwPropertyKey::Affinity`

This pulls in work from #159 to remove a now-unknown property.

Co-authored-by: Rickvanderveen <rickvdv104@gmail.com>

* Remove `ElementType::Undefined`

OpenVINO 2025.1.0 deprecates `ov_element_type_e::UNDEFINED` ([#28766]),
so this change removes it in the Rust bindings. This does create a
breaking API change across minor versions of upstream OpenVINO (v2025.0
to v2025.1); we should assume from here on out that the C API is
unstable.

[#28766]: https://github.com/openvinotoolkit/openvino/pull/28766

* ci: temporarily limit the versions tested in CI

---------

Co-authored-by: Rickvanderveen <rickvdv104@gmail.com>